move article content fetching into fewer functions
