+++ apple/OmnivoreKit/Sources/Services/DataService/Queries/LibraryItemsQuery.swift
@@ -1,4 +1,3 @@
-import Combine
 import CoreData
 import Foundation
 import Models
